refactor(settings_ui): lift restart/settings-info/feature-flag handlers to _handlers_server.py
Decomposes _save_feature_flags (C901 26 -> under) via _validate_feature_flag_batch/
_coerce_flag_value/_save_feature_flags_addon/_write_feature_flag_overrides_file, and
_restart_addon (12 -> under) via _restart_embedded/_parse_restart_slug/_restart_sibling_addon.
Moves process-identity globals + _reject_child_flags_without_parent here (re-exported for
tools_bug_report). Repoints the settings_ui.httpx.AsyncClient patch target.
