+++ src/ha_mcp/settings_ui/__init__.py
@@ -16,17 +16,14 @@ import html
 import json
 import logging
 import os
-import time
-import uuid
-from collections.abc import Awaitable, Callable, Container
+from collections.abc import Awaitable, Callable
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-import httpx
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, JSONResponse, Response
 
-from .._version import get_version, is_embedded, is_running_in_addon
+from .._version import is_running_in_addon
 from ..backup_manager import get_backup_manager
 from ..config import (
     BACKUP_OVERRIDE_FIELDS,
@@ -37,9 +34,14 @@ from ..config import (
 from ..errors import ErrorCode, create_error_response
 from ..transforms import DEFAULT_PINNED_TOOLS
 from ..utils.data_paths import get_data_dir
-from . import _persistence, _supervisor
+from . import _handlers_server, _persistence, _supervisor
 from ._handlers_backups import build_backups_handlers
 from ._handlers_fs import build_fs_handlers
+from ._handlers_server import (
+    _PROCESS_INSTANCE_ID,
+    _PROCESS_STARTED_AT,
+    build_server_handlers,
+)
 from ._handlers_theme import build_theme_handlers
 from ._handlers_tools import build_tools_handlers
 from ._persistence import (
@@ -80,17 +82,6 @@ if TYPE_CHECKING:
     from ..config import Settings
     from ..server import HomeAssistantSmartMCPServer
 
-
-# Per-process identity surfaced via ``/api/settings/info`` so the
-# restart UI can tell whether the addon actually restarted (vs. the
-# poll-cycle succeeding against the still-running OLD instance because
-# the supervisor restart silently no-op'd). Generated once at module
-# import; a fresh Python process gets a fresh value, so any restart
-# that actually swaps processes flips both. ``started_at`` is Unix
-# epoch seconds for human debuggability; ``instance_id`` is the
-# load-bearing identifier the JS poll compares against.
-_PROCESS_INSTANCE_ID: str = uuid.uuid4().hex
-_PROCESS_STARTED_AT: float = time.time()
 
 logger = logging.getLogger(__name__)
 
@@ -350,49 +341,6 @@ def _render_settings_html() -> str:
     return rendered
 
 
-def _reject_child_flags_without_parent(
-    raw_flags: dict[str, Any],
-    parent_field: str,
-    child_fields: Container[str],
-    message: Callable[[list[str]], str],
-    suggestions: list[str],
-) -> JSONResponse | None:
-    """Reject a feature-flag save that enables child flags while their
-    parent stays off after the merge.
-
-    A child flag is only valid when ``parent_field`` is truthy AFTER the
-    merge. The post-merge parent is derived from the payload (if present),
-    else the live ``Settings`` value — the same value the runtime gate
-    will see. A child turned truthy against an
-    off parent would be forced back off at runtime, so reject it now
-    rather than let the user learn the save was a no-op at next startup.
-    Turning the parent off alone is NOT rejected: children absent from the
-    payload keep their persisted value and the runtime gate handles them.
-
-    Returns a 409 ``JSONResponse`` — ``message`` receives every offending
-    child, which is also echoed in ``context["rejected"]`` — or ``None``
-    when there is nothing to reject.
-    """
-    rejected = [k for k in raw_flags if k in child_fields and bool(raw_flags[k])]
-    effective_parent = bool(
-        raw_flags.get(
-            parent_field,
-            getattr(get_global_settings(), parent_field),
-        )
-    )
-    if rejected and not effective_parent:
-        return JSONResponse(
-            create_error_response(
-                ErrorCode.VALIDATION_INVALID_PARAMETER,
-                message(rejected),
-                suggestions=suggestions,
-                context={"rejected": rejected},
-            ),
-            status_code=409,
-        )
-    return None
-
-
 def build_settings_handlers(
     server: HomeAssistantSmartMCPServer | None,
     *,
@@ -429,772 +377,6 @@ def build_settings_handlers(
     async def _settings_page(_: Request) -> HTMLResponse:
         return HTMLResponse(_render_settings_html())
 
-    async def _restart_addon(request: Request) -> JSONResponse:
-        # Embedded (in-process custom-component server): "restart" means
-        # reloading the ha_mcp_tools server config entry. Checked before the
-        # Supervisor path — the HA core container carries SUPERVISOR_TOKEN,
-        # but the Supervisor add-on API is not ours to call from here, and
-        # post-reload the new worker imports the freshly installed code
-        # (the manager purges the module cache per start).
-        if server is not None and is_embedded():
-            from ..tools.tools_dev import (
-                abort_options_flow_quietly,
-                find_server_config_entry,
-                schedule_deferred_entry_reload,
-            )
-
-            try:
-                found = await find_server_config_entry(server.client)
-            except Exception as exc:
-                # A discovery FAILURE is not "entry not found" — masking a
-                # WS/connection hiccup as not-found steers users toward
-                # reinstalling a running component. Also: the restart JS
-                # treats any 5xx as "restart in flight" and starts its
-                # poll-reload cycle, so a restart that was never initiated
-                # must answer BELOW 500.
-                logger.warning("Embedded restart: entry discovery failed: %s", exc)
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.CONNECTION_FAILED,
-                        "Could not reach Home Assistant to locate the "
-                        f"in-process server's config entry: {exc}",
-                        suggestions=[
-                            "Retry once Home Assistant is responsive",
-                        ],
-                    ),
-                    status_code=409,
-                )
-            if found is None:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        "Could not locate the in-process server's config "
-                        "entry to reload",
-                        suggestions=[
-                            "Reload the HA-MCP integration from Settings > "
-                            "Devices & Services instead",
-                        ],
-                    ),
-                    # Below 500 on purpose: no restart was initiated, and the
-                    # restart JS interprets 5xx as restart-in-flight.
-                    status_code=409,
-                )
-            entry_id, flow, _options = found
-            await abort_options_flow_quietly(server.client, flow)
-            schedule_deferred_entry_reload(server.client, entry_id)
-            return JSONResponse(
-                {"success": True, "message": "In-process server reload scheduled"}
-            )
-
-        # The sidecar process (server is None) has no Supervisor context
-        # and no live server settings — refuse cleanly. The HTTP modes
-        # that do pass a server still go through the SUPERVISOR_TOKEN
-        # check below, preserving the original behavior.
-        if server is None or not os.environ.get("SUPERVISOR_TOKEN"):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.CONFIG_VALIDATION_FAILED,
-                    "Restart only available when running as an App (add-on)",
-                    details="SUPERVISOR_TOKEN environment variable is not set",
-                ),
-                status_code=400,
-            )
-        # Optional slug from the request body lets callers restart a sibling
-        # addon instead of self. The UI's restart button posts an empty body
-        # and gets the historical self-restart behavior; the inaddon E2E
-        # suite uses ``slug`` to exercise the Supervisor restart wire
-        # contract against a non-test-critical addon (the dev addon's
-        # session would otherwise drop). The token's hassio_role gates
-        # whether the call actually succeeds for non-self targets.
-        #
-        # The slug is interpolated into the Supervisor endpoint URL, so it
-        # must be tightly constrained — Supervisor addon slugs are
-        # ``[a-z0-9_]+`` per the addon-config schema, but defending against
-        # path-traversal (``..``, ``/``, URL-encoded variants) at the edge
-        # is cheaper than relying on Supervisor to reject every bad shape.
-        # Reject anything outside ``[A-Za-z0-9_-]`` and silently fall back
-        # to ``self`` — same outcome as no body.
-        target_slug = "self"
-        try:
-            payload = await request.json()
-        except (ValueError, json.JSONDecodeError):
-            payload = None
-        if isinstance(payload, dict):
-            requested = payload.get("slug")
-            if (
-                isinstance(requested, str)
-                and requested.strip()
-                and all(c.isalnum() or c in "_-" for c in requested.strip())
-            ):
-                target_slug = requested.strip()
-
-        # Self-restart races the response flush: supervisor kills the addon
-        # mid-response, ingress sees the upstream drop, and converts it into
-        # a 5xx Bad Gateway at the browser — which the JS shows as
-        # "Restart failed" even though the restart actually succeeded.
-        # Schedule the supervisor POST from a background task so the JSON
-        # response below flushes BEFORE supervisor can kill us.
-        #
-        # Non-self slugs target a sibling addon: the inaddon E2E suite
-        # exercises that path against a non-test-critical addon, and we
-        # want the supervisor response (including 4xx) to surface
-        # synchronously so the test can assert on it. Keep the original
-        # synchronous behavior for that path.
-        if target_slug == "self":
-            _supervisor._schedule_supervisor_self_restart(server.settings.verify_ssl)
-            return JSONResponse({"success": True, "message": "Restart initiated"})
-
-        endpoint = f"/addons/{target_slug}/restart"
-        try:
-            async with _supervisor.make_supervisor_httpx_client(
-                timeout=5.0, verify=server.settings.verify_ssl
-            ) as client:
-                resp = await client.post(endpoint)
-        except (httpx.ReadError, httpx.RemoteProtocolError):
-            # Connection dropped mid-request — restart is happening.
-            # `ConnectError` is deliberately NOT in this tuple: it fires
-            # before a connection is established (DNS failure, TCP refused,
-            # Supervisor socket misconfigured) and means the restart was
-            # never initiated. Falls through to the `httpx.HTTPError`
-            # handler below, which returns 502 + CONNECTION_FAILED.
-            logger.info(
-                "Restart request connection dropped (expected during self-restart)"
-            )
-            return JSONResponse({"success": True, "message": "Restart initiated"})
-        except httpx.HTTPError as e:
-            logger.exception("Failed to reach Supervisor for restart")
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.CONNECTION_FAILED,
-                    f"Failed to reach Supervisor: {e}",
-                ),
-                status_code=502,
-            )
-
-        if resp.status_code >= 400:
-            body = resp.text
-            logger.error(
-                "Supervisor restart failed (slug=%s): %d %s",
-                target_slug,
-                resp.status_code,
-                body,
-            )
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.INTERNAL_ERROR,
-                    f"Supervisor returned {resp.status_code}: {body[:500]}",
-                ),
-                status_code=502,
-            )
-        return JSONResponse({"success": True, "message": "Restart initiated"})
-
-    async def _settings_info(_: Request) -> JSONResponse:
-        # Sidecar is never the add-on entrypoint regardless of inherited
-        # SUPERVISOR_TOKEN — see docstring above for the broken-button
-        # rationale. ``is_sidecar`` flag drives the in-page Stop Sidecar
-        # button (HTML show/hide); it MUST NOT leak True for HTTP modes
-        # since stopping the FastMCP-mounted route would mean killing
-        # the MCP server itself.
-        #
-        # ``instance_id`` + ``started_at`` are surfaced so the
-        # restart-then-reload JS cycle can prove a restart actually
-        # happened (the value flips across processes) instead of
-        # trusting that a poll returning 200 means the new instance
-        # is up — a silent restart no-op would otherwise see the OLD
-        # instance still answering and the page would reload to the
-        # same state.
-        addon = False if is_sidecar else is_running_in_addon()
-        # ``version`` surfaces the running ha-mcp version in the UI
-        # footer. ``get_version`` handles the env-override → package
-        # metadata → fallback chain itself, so a read per request is
-        # cheap. ``HA_MCP_BUILD_VERSION`` is set by both stable and
-        # dev addon Dockerfiles, so the version reads correctly on
-        # either channel; standalone Docker / uvx falls back to the
-        # installed package's metadata.
-        try:
-            # Executor: get_version's distribution-ownership scan reads
-            # metadata for every installed package — too heavy for the
-            # event loop on an endpoint the restart cycle polls.
-            version = await asyncio.to_thread(get_version)
-        except Exception:  # pragma: no cover — defensive only
-            logger.warning("get_version() raised; omitting version from info")
-            version = None
-        # ``deployment_mode`` reuses the bug-report detector so the UI and
-        # ha_report_issue can never disagree about where the server runs —
-        # the embedded (in-process custom component) server would otherwise
-        # read as plain "docker" here (/.dockerenv exists in the HA core
-        # container).
-        from ..tools.tools_bug_report import _detect_installation_method
-
-        return JSONResponse(
-            {
-                "is_addon": addon,
-                "is_sidecar": is_sidecar,
-                "deployment_mode": (
-                    "sidecar" if is_sidecar else _detect_installation_method()
-                ),
-                "instance_id": _PROCESS_INSTANCE_ID,
-                "started_at": _PROCESS_STARTED_AT,
-                "version": version,
-            }
-        )
-
-    async def _live_addon_options() -> dict[str, Any]:
-        """Best-effort fetch of the add-on's current ``/data/options.json``.
-
-        Read-consistency with the add-on Configuration tab: the GET handlers
-        below otherwise display boot-time env values, so a user who edits the
-        add-on config (or saves from this web UI) WITHOUT restarting would see
-        stale values. For add-on-synced / schema-backed fields the live value
-        lives in Supervisor's options, so surface that — the latest SAVED
-        value, which is what becomes active after restart (the existing
-        "Restart required after save" banner conveys apply-on-restart).
-
-        Returns an empty dict outside add-on mode, without a live server, or
-        on any fetch failure, so callers fall back to the boot-env value and
-        the page never crashes or blanks a field.
-        """
-        if not is_running_in_addon() or server is None:
-            return {}
-        try:
-            options, err = await _supervisor._supervisor_fetch_current_options(
-                server.settings.verify_ssl
-            )
-        except Exception as exc:  # pragma: no cover - defensive belt-and-braces
-            logger.debug(
-                "Live add-on options fetch raised %s — using boot-env values",
-                exc,
-            )
-            return {}
-        if err is not None:
-            logger.debug(
-                "Live add-on options fetch failed (%s): %s — using boot-env values",
-                err.kind,
-                err.message,
-            )
-            return {}
-        return options
-
-    async def _get_feature_flags(_: Request) -> JSONResponse:
-        """Return live feature-flag values + per-field origin + editable flag.
-
-        Per-field origin/editable matrix (see
-        :func:`config.get_feature_flag_origin`):
-
-        - ``"addon"``: editable — POST routes through Supervisor
-          ``/addons/self/options`` and triggers a restart so
-          ``start.py`` writes the new env vars at next boot. The
-          add-on Configuration tab and this web UI now share state
-          bidirectionally; changing a toggle in either surface is
-          reflected in the other after the addon restart.
-        - ``"env"``: read-only — env var explicitly set wins; user
-          must unset it to edit here.
-        - ``"file"`` / ``"default"``: editable — POST writes the
-          override file in place.
-
-        The envelope shape (``flags`` dict of
-        ``{value, origin, editable, type, env_var, min?, max?}``
-        entries) is intentionally generic so other settings surfaces
-        can render rows with the same JS code.
-        """
-        from ..config import (
-            _FEATURE_FLAG_INT_BOUNDS,
-            FEATURE_FLAG_FIELDS,
-            get_feature_flag_origin,
-            get_global_settings,
-        )
-
-        settings = get_global_settings()
-        # Read-consistency with the add-on Configuration tab: a flag the user
-        # saved in the Config tab is already in Supervisor's live options, but
-        # the boot-env value goes stale until a restart. Surface the latest
-        # SAVED value for any flag present in live_options (see
-        # _live_addon_options) — not just origin=="addon" rows. On a dev add-on
-        # a flag saved before restart has origin "default"/"file" (its env var
-        # wasn't written at this boot), yet the fresh value is the one to show.
-        # Only origin=="env" stays pinned (an explicit env var the user must
-        # unset to change), and live_options is {} outside add-on mode so this
-        # is a no-op standalone.
-        live_options = await _live_addon_options()
-        flags: dict[str, Any] = {}
-        for field_name, env_name, ftype in FEATURE_FLAG_FIELDS:
-            origin = get_feature_flag_origin(env_name)
-            value = getattr(settings, field_name)
-            if field_name in live_options and origin != "env":
-                value = live_options[field_name]
-            entry: dict[str, Any] = {
-                "value": value,
-                "origin": origin,
-                "editable": origin in ("addon", "file", "default"),
-                "type": ftype.__name__,
-                "env_var": env_name,
-            }
-            if ftype is int:
-                bounds = _FEATURE_FLAG_INT_BOUNDS.get(field_name)
-                if bounds is not None:
-                    entry["min"], entry["max"] = bounds
-            flags[field_name] = entry
-        from ..config import BETA_FEATURE_FIELDS
-
-        return JSONResponse(
-            {
-                "flags": flags,
-                "beta_sub_flags": list(BETA_FEATURE_FIELDS),
-                # Drives addon-aware locked-banner copy in the JS —
-                # "unset env var" is misleading where HA Supervisor
-                # owns the env.
-                "is_addon": is_running_in_addon(),
-            }
-        )
-
-    async def _save_feature_flags(request: Request) -> JSONResponse:
-        """Persist UI-edited feature-flag values.
-
-        Routing by per-field origin (see
-        :func:`config.get_feature_flag_origin`):
-
-        - **addon**: POST the merged options to Supervisor and return
-          ``restart_required=True``. ``start.py`` will re-derive env
-          vars from ``config.yaml`` on the next addon boot — but the
-          actual restart is fired by the user clicking the global
-          Restart Add-on button, NOT by this handler. Web UI edits
-          and Configuration-tab edits land in the same place, so the
-          two surfaces stay in sync after the restart.
-        - **env**: refuse — env var explicitly set wins. Returns
-          ``VALIDATION_INVALID_PARAMETER`` with the env var name so
-          the UI can surface the locking source.
-        - **file** / **default**: merge into the override file in the
-          data dir; takes effect on the next
-          ``get_global_settings()`` call (cache reset). The response
-          still carries ``restart_required=True`` because most
-          flag descriptions advertise "Requires restart to take
-          effect" — the UI shows the banner regardless of mode.
-        """
-        from ..config import (
-            _FEATURE_FLAG_INT_BOUNDS,
-            _FEATURE_FLAG_OVERRIDE_FILENAME,
-            FEATURE_FLAG_FIELDS,
-            get_feature_flag_origin,
-        )
-        from ..utils.data_paths import get_data_dir
-
-        try:
-            body = await request.json()
-        except (ValueError, TypeError):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_JSON,
-                    "Invalid JSON body",
-                ),
-                status_code=400,
-            )
-        if not isinstance(body, dict):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_PARAMETER,
-                    "Request body must be a JSON object",
-                ),
-                status_code=400,
-            )
-        raw_flags = body.get("flags", {})
-        if not isinstance(raw_flags, dict):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_PARAMETER,
-                    "'flags' must be an object mapping field names to values",
-                ),
-                status_code=400,
-            )
-
-        # Master beta-gate check: a beta sub-flag may only be enabled
-        # when the master ``enable_beta_features`` is on AFTER the merge
-        # (see ``_reject_child_flags_without_parent`` for the post-merge
-        # derivation and why a parent-off-only save is not rejected). All
-        # offending sub-flags are listed in the rejection.
-        #
-        # Applied in BOTH standalone and addon mode.
-        # The earlier "skip in addon mode" carve-out existed because
-        # start.py used to auto-write ENABLE_BETA_FEATURES=true from
-        # any beta sub-flag presence; that path is now demoted to a
-        # one-cycle legacy bridge. On dev addon, start.py writes the
-        # master env var from the schema-bound options key. On stable
-        # addon, the master is not in schema and the standalone
-        # web-UI master path remains the gate (the gate read
-        # falls through to the override-file value). Either way the
-        # gate is sound to apply uniformly.
-        from ..config import (
-            BETA_FEATURE_FIELDS as _BETA_SUB,
-        )
-
-        beta_rejection = _reject_child_flags_without_parent(
-            raw_flags,
-            "enable_beta_features",
-            _BETA_SUB,
-            lambda rejected: (
-                "Cannot enable beta sub-flag(s) "
-                f"{', '.join(rejected)} while the master "
-                "'Enable beta features' toggle is off. Include "
-                "enable_beta_features=true in the same save, or "
-                "flip the master on first."
-            ),
-            [
-                "Include enable_beta_features=true in the same save "
-                + "payload as the sub-flag(s).",
-                "Or turn on the master 'Enable beta features' toggle "
-                + "first, then enable the sub-flag(s).",
-            ],
-        )
-        if beta_rejection is not None:
-            return beta_rejection
-
-        # Strict-mandatory-BPS dependency gate: strict mode
-        # (``enable_strict_mandatory_bps``) is a child of
-        # ``enable_mandatory_bps`` and is inert unless the parent is on.
-        # Same post-merge derivation and no-parent-off-rejection as the
-        # beta gate above (see ``_reject_child_flags_without_parent``).
-        strict_rejection = _reject_child_flags_without_parent(
-            raw_flags,
-            "enable_mandatory_bps",
-            ("enable_strict_mandatory_bps",),
-            lambda _rejected: (
-                "Cannot enable strict best-practices mode "
-                "('enable_strict_mandatory_bps') while the parent "
-                "'Attach best-practice skills on writes' "
-                "(enable_mandatory_bps) toggle is off. Strict mode is "
-                "a child of that toggle and has no effect without it. "
-                "Include enable_mandatory_bps=true in the same save, or "
-                "turn the parent on first."
-            ),
-            [
-                "Include enable_mandatory_bps=true in the same save "
-                + "payload as enable_strict_mandatory_bps.",
-                "Or turn on the parent 'Attach best-practice skills on "
-                + "writes' toggle first, then enable strict mode.",
-            ],
-        )
-        if strict_rejection is not None:
-            return strict_rejection
-
-        # Build the validated override dict. Reject unknown fields and
-        # env-locked fields up front so the user gets a precise error
-        # instead of a silent no-op. ``addon``-origin fields are now
-        # editable — they route through Supervisor below.
-        known: dict[str, tuple[str, type]] = {
-            fname: (ename, ftype) for fname, ename, ftype in FEATURE_FLAG_FIELDS
-        }
-        new_overrides: dict[str, Any] = {}
-        # Per ``config.get_feature_flag_origin``, addon mode (i.e.
-        # ``SUPERVISOR_TOKEN`` set) makes the helper return ``"addon"``
-        # for *every* registered flag — there is no path that yields a
-        # mixed addon + file/default batch from a single addon-mode UI
-        # session. The loop still tracks ``addon_writes`` per-field as
-        # belt-and-braces in case a future origin-resolution change
-        # breaks that invariant; the post-loop assertion below makes the
-        # invariant explicit so a regression fails loudly instead of
-        # silently routing a file-mode write through Supervisor.
-        addon_writes = False
-        file_or_default_writes = False
-        for field_name, raw in raw_flags.items():
-            if field_name not in known:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"Unknown feature flag: {field_name!r}",
-                    ),
-                    status_code=400,
-                )
-            env_name, ftype = known[field_name]
-            origin = get_feature_flag_origin(env_name)
-            if origin == "addon":
-                addon_writes = True
-            elif origin in ("file", "default"):
-                file_or_default_writes = True
-            else:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        (
-                            f"{field_name!r} is locked by {origin}. "
-                            f"Adjust the {env_name} env var "
-                            "(or App (add-on) configuration) instead."
-                        ),
-                    ),
-                    status_code=400,
-                )
-            if ftype is bool:
-                if not isinstance(raw, bool):
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            f"{field_name!r} must be a boolean",
-                        ),
-                        status_code=400,
-                    )
-                new_overrides[field_name] = bool(raw)
-            elif ftype is int:
-                if isinstance(raw, bool) or not isinstance(raw, int):
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            f"{field_name!r} must be an integer",
-                        ),
-                        status_code=400,
-                    )
-                bounds = _FEATURE_FLAG_INT_BOUNDS.get(field_name)
-                if bounds is not None and not bounds[0] <= raw <= bounds[1]:
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            (
-                                f"{field_name!r} must be between "
-                                f"{bounds[0]} and {bounds[1]}"
-                            ),
-                        ),
-                        status_code=400,
-                    )
-                new_overrides[field_name] = int(raw)
-            else:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        (f"{field_name!r} has an unsupported type for UI editing"),
-                    ),
-                    status_code=400,
-                )
-
-        # Master-off no longer cascades into sub-flag values. The
-        # runtime master gate in ``_apply_feature_flag_overrides``
-        # continues to force every
-        # beta sub-flag to False whenever the master is off, so the
-        # tools stay disabled at runtime regardless of file state.
-        # Leaving the sub-flag values in the override file means
-        # re-enabling the master restores the user's prior sub-flag
-        # selections automatically — without it the user had to
-        # re-check each sub-flag individually after every
-        # master-off / master-on cycle, which is the wrong UX trade
-        # for an opt-in beta surface.
-        #
-        # The master-gate check above still rejects payloads that try
-        # to enable a sub-flag while the effective master is off, so
-        # users can't land a "sub=true while master=false in same
-        # payload" inconsistency. Pre-existing sub-flag truthy values
-        # in the file are kept verbatim.
-
-        # Addon-mode writes go to Supervisor instead of the override file:
-        # ``start.py`` reads ``config.yaml`` options on every boot and
-        # writes the env vars that Settings consumes, so the override
-        # file is ignored anyway (see config.get_feature_flag_origin).
-        # POST a merged options dict (current options + this delta) so
-        # required schema keys like ``backup_hint`` survive — see
-        # _supervisor_merge_and_post_options for the rationale, and the
-        # parallel handling in _save_backup_config.
-        #
-        # Reject mixed-origin batches loudly. The current
-        # ``get_feature_flag_origin`` implementation guarantees a single
-        # mode per request, but if a future change broke that invariant
-        # we would silently route a file/default-origin field through
-        # Supervisor (which would reject it as an unknown schema key in
-        # production addon mode) — better to fail clearly here.
-        if addon_writes and file_or_default_writes:
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.INTERNAL_ERROR,
-                    (
-                        "Batch contains a mix of addon-origin and "
-                        "file/default-origin fields; route each batch "
-                        "through a single persistence path."
-                    ),
-                ),
-                status_code=500,
-            )
-        if addon_writes:
-            # ``addon_writes=True`` is equivalent to
-            # ``is_running_in_addon()`` (origin=="addon" requires
-            # SUPERVISOR_TOKEN per ``get_feature_flag_origin``), so the
-            # only guard we still need here is the sidecar-shape
-            # ``server is None``. The helpers below catch the missing-
-            # token ``RuntimeError`` from ``make_supervisor_httpx_client``
-            # as defense-in-depth.
-            if server is None:
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        "Feature-flag POST requires a live MCP server",
-                    ),
-                    status_code=500,
-                )
-            ok, err = await _supervisor._supervisor_merge_and_post_options(
-                server.settings.verify_ssl, new_overrides
-            )
-            if not ok:
-                if err is None:
-                    # ``ok=False`` with no error is a contract bug —
-                    # bail with INTERNAL_ERROR rather than letting an
-                    # AttributeError leak under ``python -O``.
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.INTERNAL_ERROR,
-                            "Supervisor helper returned ok=False with no error",
-                            suggestions=[
-                                "Check the Home Assistant Supervisor logs and "
-                                + "the App (add-on) logs for the underlying failure.",
-                                "Report this at "
-                                + "https://github.com/homeassistant-ai/ha-mcp/issues "
-                                + "if it persists. This indicates an internal bug.",
-                            ],
-                        ),
-                        status_code=500,
-                    )
-                logger.warning(
-                    "Supervisor feature-flag update failed (%s): %s",
-                    err.kind,
-                    err.message,
-                )
-                # Transport failures get CONNECTION_FAILED so the UI shows the
-                # "is HA reachable" suggestions; supervisor schema rejections
-                # get CONFIG_VALIDATION_FAILED with supervisor's real status
-                # code preserved so the UI shows the actual 4xx, not a generic
-                # 502. See _SupervisorOptionsError for the rationale.
-                code = (
-                    ErrorCode.CONNECTION_FAILED
-                    if err.kind == "transport"
-                    else ErrorCode.CONFIG_VALIDATION_FAILED
-                )
-                return JSONResponse(
-                    create_error_response(code, err.message),
-                    status_code=err.status_code,
-                )
-            # Unified restart flow: every save that requires an addon
-            # restart (Tools, Server Settings, Backups) returns
-            # ``restart_required=True`` and lets the user pick when to
-            # fire the actual restart via the global Restart Add-on
-            # button. Don't auto-restart from the save handler —
-            # supervisor would kill the addon before this JSON response
-            # could flush through HA ingress, surfacing as a spurious
-            # "Restart failed" alert at the browser.
-            return JSONResponse(
-                {
-                    "success": True,
-                    "applied": new_overrides,
-                    "mode": "addon",
-                    "restart_required": True,
-                }
-            )
-
-        # Merge with the existing override file so a partial POST
-        # only updates the keys it actually included — the front-end
-        # POSTs individual changes, not the entire matrix.
-        #
-        # Three failure modes for the read:
-        #
-        # * file missing → fresh dict, normal first-write path.
-        # * file unreadable (PermissionError, etc.) → 500. We can't
-        #   silently fall back to {} and overwrite, because that
-        #   would silently drop every flag the user had previously
-        #   persisted (the same data we can't read is still on
-        #   disk).
-        # * file present but corrupt JSON → 409. Same data-loss
-        #   hazard: a partial write from a prior crash, blindly
-        #   overwriting, erases anything past the corruption point.
-        #   Return a clear error so the user can inspect / delete
-        #   the file manually.
-        path = get_data_dir() / _FEATURE_FLAG_OVERRIDE_FILENAME
-        # Serialise concurrent saves on the shared override file so
-        # two overlapping requests can't interleave their RMW. Lock is
-        # held only for the read+merge+atomic-write window; pure
-        # validation above does not need it.
-        async with _persistence._get_override_file_lock():
-            existing: dict[str, Any] = {}
-            try:
-                existing_raw = path.read_text()
-            except FileNotFoundError:
-                existing_raw = None
-            except OSError as exc:
-                logger.warning("Cannot read %s", path, exc_info=True)
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        (
-                            f"Could not read existing feature flags "
-                            f"({type(exc).__name__}: {exc}); refusing to "
-                            "overwrite to preserve prior toggles. "
-                            "Check filesystem permissions and retry."
-                        ),
-                    ),
-                    status_code=500,
-                )
-            if existing_raw is not None:
-                try:
-                    parsed = json.loads(existing_raw)
-                except json.JSONDecodeError as exc:
-                    logger.warning(
-                        "Existing %s is corrupt: %s", path, exc, exc_info=True
-                    )
-                    return JSONResponse(
-                        create_error_response(
-                            ErrorCode.VALIDATION_INVALID_PARAMETER,
-                            (
-                                f"Existing override file at {path} is not "
-                                f"valid JSON ({exc}); refusing to overwrite "
-                                "to preserve prior toggles. Inspect or "
-                                "delete the file manually and retry."
-                            ),
-                        ),
-                        status_code=409,
-                    )
-                if isinstance(parsed, dict):
-                    existing = parsed
-                # else: non-dict JSON (list, scalar) — treat as empty;
-                # we're about to write a dict either way and there's
-                # no prior toggle state to preserve from a non-object
-                # root.
-            existing.update(new_overrides)
-
-            # Atomic write: tmp + rename. ``path.write_text`` is
-            # O_TRUNC + write — a crash mid-write leaves an empty /
-            # truncated file that the next
-            # ``_read_feature_flag_override_file`` call would refuse,
-            # losing every prior toggle.
-            tmp = path.with_suffix(path.suffix + ".tmp")
-            try:
-                tmp.write_text(json.dumps(existing, indent=2))
-                os.replace(tmp, path)
-            except OSError as exc:
-                logger.warning("Could not write %s", path, exc_info=True)
-                with contextlib.suppress(FileNotFoundError, OSError):
-                    tmp.unlink()
-                return JSONResponse(
-                    create_error_response(
-                        ErrorCode.INTERNAL_ERROR,
-                        f"Could not persist feature flags: {exc}",
-                    ),
-                    status_code=500,
-                )
-
-        # Publish the change so the same process picks it up on the
-        # next ``get_global_settings()`` call. The cached singleton
-        # must not return the stale pre-write values to subsequent
-        # /api/settings/features GETs.
-        #
-        # ``restart_required=True`` because the feature flags here gate
-        # tool registration, FastMCP transforms, and other startup-time
-        # reads. File-mode persists the value, but the live process
-        # keeps the old behavior until the MCP host is restarted
-        # (Claude Desktop relaunch, Docker container restart, etc.).
-        # Surfacing the banner is the same contract Tools, Server
-        # Settings, and Backups all advertise.
-        _reset_global_settings()
-        return JSONResponse(
-            {
-                "success": True,
-                "applied": new_overrides,
-                "mode": "file",
-                "restart_required": True,
-            }
-        )
-
-    # ---- Auto-backup routes (#1288) ----
-
     async def _get_advanced_settings(_: Request) -> JSONResponse:
         """Return advanced (non-feature-flag, non-backup) settings + per-field
         origin + editable flag.
@@ -1215,7 +397,6 @@ def build_settings_handlers(
             ADVANCED_SETTINGS_FIELDS,
             OAUTH_MODE_TOKEN,
             _read_feature_flag_override_file,
-            get_global_settings,
         )
 
         settings = get_global_settings()
@@ -1227,7 +408,7 @@ def build_settings_handlers(
         # fields (origin "addon") live in /data/options.json, so the boot-env
         # value goes stale after a config edit without a restart. Surface the
         # latest SAVED options value for those (see _live_addon_options).
-        live_options = await _live_addon_options()
+        live_options = await _handlers_server._live_addon_options(server)
         fields: list[dict[str, Any]] = []
         for (
             fname,
@@ -1698,10 +879,6 @@ def build_settings_handlers(
     handlers: dict[str, Any] = {
         "root_page": _root_page,
         "settings_page": _settings_page,
-        "restart_addon": _restart_addon,
-        "settings_info": _settings_info,
-        "get_feature_flags": _get_feature_flags,
-        "save_feature_flags": _save_feature_flags,
         "get_advanced_settings": _get_advanced_settings,
         "save_advanced_settings": _save_advanced_settings,
     }
@@ -1733,6 +910,7 @@ def build_settings_handlers(
     handlers.update(build_fs_handlers(server))
     handlers.update(build_tools_handlers(server))
     handlers.update(build_backups_handlers(server))
+    handlers.update(build_server_handlers(server, is_sidecar=is_sidecar))
 
     return handlers
