+++ tests/src/unit/test_settings_ui.py
@@ -1033,7 +1033,7 @@ class TestRestartAddon:
         cm.__aenter__ = AsyncMock(return_value=mock_client)
         cm.__aexit__ = AsyncMock(return_value=None)
 
-        with patch("ha_mcp.settings_ui.httpx.AsyncClient", return_value=cm):
+        with patch("ha_mcp.settings_ui._handlers_server.httpx.AsyncClient", return_value=cm):
             resp = await restart(request)
 
         assert resp.status_code == 200
